moved addtocard function to productDetails
