+++ src/js/product.js
@@ -1,4 +1,4 @@
-import { setLocalStorage, getParam } from "./utils.mjs";
+import { getParam } from "./utils.mjs";
 import { findProductById } from "./productData.mjs";
 
 const productId = getParam("product");
@@ -7,16 +7,3 @@ const productId = getParam("product");
 console.log(findProductById(productId));
 
 
-function addProductToCart(product) {
-  setLocalStorage("so-cart", product);
-}
-// add to cart button event handler
-async function addToCartHandler(e) {
-  const product = await findProductById(e.target.dataset.id);
-  addProductToCart(product);
-}
-
-// add listener to Add to Cart button
-document
-  .getElementById("addToCart")
-  .addEventListener("click", addToCartHandler);
