+++ src/js/productDetails.mjs
@@ -1,0 +1,23 @@
+
+
+
+
+
+import { setLocalStorage } from "./utils.mjs";
+
+
+// 2. addToCart 
+function addProductToCart(product) {
+    setLocalStorage("so-cart", product);
+  }
+  // add to cart button event handler
+  async function addToCartHandler(e) {
+    const product = await findProductById(e.target.dataset.id);
+    addProductToCart(product);
+  }
+  
+  // add listener to Add to Cart button
+  document
+    .getElementById("addToCart")
+    .addEventListener("click", addToCartHandler);
+
